refactor(web): 탐색, 좋아요 Frame기반 PortfolioList 컴포넌트 교체
